+++ capywfa/capywfa.py
@@ -91,6 +91,9 @@ def pass1_map_bom(bom, sw360_url, sw360_token):
         item for item in result.components
         if get_cdx(item, "MapResult") not in (
             MapResult.MATCH_BY_NAME, MapResult.SIMILAR_COMPONENT_FOUND)]
+    for item in result.components:
+        if "qualifiers-ignored" in get_cdx(item, "MapResultById").split():
+            set_cdx(item, "Sw360SourceFileCheck", "force-content-check")
     return result
 
 
@@ -98,7 +101,7 @@ def pass3_download_sources(bom):
     for item in bom.components:
         if not (get_cdx(item, "MapResult") == MapResult.NO_MATCH
                 or (MapBom.is_good_match(get_cdx(item, "MapResult"))
-                    and get_cdx(item, "Sw360SourceFileChecked") != "true")):
+                    and get_cdx(item, "Sw360SourceFileCheck") != "passed")):
             set_cdx(item, "SourceFileDownload", "skip")
     return bom
 
@@ -179,7 +182,7 @@ def pass6_link_releases(bom, sw360_url, sw360_token, project_id, minus_id):
         if get_cdx(item, "Sw360Id")
         and (get_cdx(item, "MapResult") == MapResult.NO_MATCH
              or MapBom.is_good_match(get_cdx(item, "MapResult")))
-        and get_cdx(item, "Sw360SourceFileChecked") == "true"]
+        and get_cdx(item, "Sw360SourceFileCheck") == "passed"]
     print("Found", len(todo_bom.components), "releases to link.")
 
     if minus_id is not None:
@@ -303,6 +306,9 @@ def main():
     print("Release matches in SW360:",
           len([item for item in bom.components
                if MapBom.is_good_match(get_cdx(item, "MapResult"))]))
+    print("Releases marked for source file check (qualifiers didn't match):",
+          len([item for item in bom.components
+               if get_cdx(item, "Sw360SourceFileCheck") == "force-content-check"]))
     print("Component (purl) matches in SW360:",
           len([item for item in bom.components
                if get_cdx(item, "MapResult") == MapResult.NO_MATCH
@@ -316,7 +322,8 @@ def main():
     print("== Pass 2: Verify SW360 sources (quick) ==")
     print()
 
-    if get_cdx(bom.components[0], "Sw360SourceFileChecked") and not args.remap:
+    if (get_cdx(bom.components[0], "Sw360SourceFileCheck") != "force-content-check"
+            and not args.remap):
         print("BOM seems to contain check results, skipping...")
     else:
         # without pkg_dir, it will only verify SW360's attachment `checkStatus`
@@ -328,7 +335,7 @@ def main():
     print()
     print("Verified sources:",
           len([item for item in bom.components
-              if get_cdx(item, "Sw360SourceFileChecked") == "true"]))
+              if get_cdx(item, "Sw360SourceFileCheck") == "passed"]))
 
     print()
     print("== Pass 3: Download missing and unchecked sources ==")
@@ -378,7 +385,7 @@ def main():
     missing_verification_count = len(
         [item for item in bom.components
          if MapBom.is_good_match(get_cdx(item, "MapResult"))
-         and get_cdx(item, "Sw360SourceFileChecked") != "true"])
+         and get_cdx(item, "Sw360SourceFileCheck") != "passed"])
 
     if missing_verification_count == 0:
         print("Verified all existing sources.")
@@ -401,11 +408,11 @@ def main():
     print()
     print("Verified sources:",
           len([item for item in bom.components
-              if get_cdx(item, "Sw360SourceFileChecked") == "true"]))
+              if get_cdx(item, "Sw360SourceFileCheck") == "passed"]))
     print("Invalid sources:",
           len([item for item in bom.components
               if MapBom.is_good_match(get_cdx(item, "MapResult"))
-              and get_cdx(item, "Sw360SourceFileChecked") != "true"]))
+              and get_cdx(item, "Sw360SourceFileCheck") != "passed"]))
 
     print()
     print("== Pass 6: Link releases ==")
@@ -487,7 +494,7 @@ def main():
 
     todo_unchecked = [item for item in bom.components
                       if MapBom.is_good_match(get_cdx(item, "MapResult"))
-                      and get_cdx(item, "Sw360SourceFileChecked") != "true"
+                      and get_cdx(item, "Sw360SourceFileCheck") != "passed"
                       and get_cdx(item, "SourceFileDownload") != "failed"]
     if len(todo_unchecked) > 0:
         print(len(todo_unchecked), "SW360 sources couldn't be verified.")
@@ -497,6 +504,23 @@ def main():
             """))
         print("BOM items with attachment differences:")
         for item in todo_unchecked:
+            print("-", item.name, item.version,
+                  args.url + "/group/guest/components/-/component/release/"
+                  "detailRelease/"
+                  + get_cdx(item, "Sw360Id")+"#/tab-Attachments")
+        print()
+        problem = True
+
+    todo_checked_with_qualifier_mismatch = [
+        item for item in bom.components
+        if MapBom.is_good_match(get_cdx(item, "MapResult"))
+        and "qualifiers-ignored" in get_cdx(item, "MapResultById").split()
+        and get_cdx(item, "Sw360SourceFileCheck") == "passed"]
+    if len(todo_checked_with_qualifier_mismatch) > 0:
+        print(len(todo_checked_with_qualifier_mismatch),
+              "releases with qualifier mismatch, but source file check passed.")
+        print("Releases with correct sources but deviating PURL qualifiers:")
+        for item in todo_checked_with_qualifier_mismatch:
             print("-", item.name, item.version,
                   args.url + "/group/guest/components/-/component/release/"
                   "detailRelease/"