+++ capywfa/verify_sources.py
@@ -77,22 +77,19 @@ def verify_sources(bom, sw360_url, sw360_token, trusted_verifiers,
 
     for item in bom.components:
         print()
-        if get_cdx(item, "Sw360SourceFileChecked") == "true":
+        source_check = get_cdx(item, "Sw360SourceFileCheck")
+        if source_check == "passed":
             print("Skipping", item.name, item.version,
                   "- already checked")
             continue
 
-        set_cdx(item, "Sw360SourceFileChecked", "false")
+        if source_check != "force-content-check":
+            set_cdx(item, "Sw360SourceFileCheck", "failed")
 
         if get_cdx(item, "MapResult") in (MapResult.MATCH_BY_NAME,
                                           MapResult.NO_MATCH):
             print("Skipping", item.name, item.version,
                   "- MapResult is", get_cdx(item, "MapResult"))
-            continue
-
-        if get_cdx(item, "SourceFileDownload") == "failed":
-            print("Skipping", item.name, item.version,
-                  "- Debian download failed")
             continue
 
         sw360id = get_cdx(item, "Sw360Id")
@@ -112,23 +109,30 @@ def verify_sources(bom, sw360_url, sw360_token, trusted_verifiers,
 
         if len(sources) != 1:
             print("ERROR:", len(sources), "sources found")
-            set_cdx(item, "Sw360SourceFileChecked", "multiple sources")
+            set_cdx(item, "Sw360SourceFileCheck", "multiple-sources")
             continue
 
         attachment_id = sources[0]['_links']['self']['href']
         attachment_id = sw360.get_id_from_href(attachment_id)
 
         print("checking", item.name, "release", sw360id,
-              "- attachment", attachment_id)
+              "- attachment", attachment_id,
+              "(source check forced)" if source_check == "force-content-check" else "")
         # print("createdBy:", sources[0]['createdBy'],
         #       "on", sources[0]['createdOn'])
         checkedby = sources[0].get('checkedBy', "")
         # print("checkStatus:", sources[0]['checkStatus'], "by", checkedby)
-        if (sources[0]['checkStatus'] == 'ACCEPTED'
+        if (source_check != "force-content-check"
+                and sources[0]['checkStatus'] == 'ACCEPTED'
                 and checkedby in trusted_verifiers):
-            set_cdx(item, "Sw360SourceFileChecked", "true")
+            set_cdx(item, "Sw360SourceFileCheck", "passed")
             print("OK: Trusted verifier", checkedby,
                   "approved on", sources[0].get('checkedOn'))
+            continue
+
+        if get_cdx(item, "SourceFileDownload") == "failed":
+            print("Skipping", item.name, item.version,
+                  "- Debian download failed")
             continue
 
         source_ext_ref = CycloneDxSupport.get_ext_ref_source_file(item)
@@ -137,7 +141,7 @@ def verify_sources(bom, sw360_url, sw360_token, trusted_verifiers,
             continue
 
         if sources[0]['sha1'] == CycloneDxSupport.get_source_file_hash(item):
-            set_cdx(item, "Sw360SourceFileChecked", "true")
+            set_cdx(item, "Sw360SourceFileCheck", "passed")
             print("OK: Hash match.")
             set_check_status(sw360, sw360id, sources[0], attachment_id)
             continue
@@ -177,11 +181,14 @@ def verify_sources(bom, sw360_url, sw360_token, trusted_verifiers,
             shell=True)
         if ret == 0:
             print("OK:", sw360_file, "identical to", our_file)
-            set_cdx(item, "Sw360SourceFileChecked", "true")
+            set_cdx(item, "Sw360SourceFileCheck", "passed")
             shutil.rmtree("verify/local-"+our_file+"-unzip")
             shutil.rmtree("verify/sw360-"+sw360_file+"-unzip")
             os.remove("verify/"+sw360_file)
             set_check_status(sw360, sw360id, sources[0], attachment_id)
+        else:
+            print("ERROR - source files differ!")
+            set_cdx(item, "Sw360SourceFileCheck", "failed")
 
     return bom